website login update
Changes that attempt an ajax request to allow a user to be identified.
IT works the user actually does get authenticated and all that's good.
However, there is an error going on somewhere and the HTTP header is
attempting to be set again producing an error. This usually indicates a
duplicate function call or inappropriate render.  The logic is correct,
though